Converted "stats" variable to a structure.
Earlier "stats.mat" would be saved as a meta variable, which could be accessed to determine properties of gene across both subpopulations. However, on some systems, saving of a file by a software was disallowed. Hence, instead of saving it as a separate variable on the hard disc, stats was passed as an argument to the function named "find_a_b.m"